+++ dev/user.clj
@@ -1,64 +1,44 @@
 (ns user
   (:require
    [clj-reload.core :as reload]
-   [java-time.api :as jt]
+   [clojure.java.io :as io]
+   [charred.api :as charred]
+   [environ.core :refer [env]]
    [taoensso.telemere :as tel]
    [hkimjp.datascript :as ds]
-   [hkimjp.jpy.system :refer [start-system stop-system]]))
+   [hkimjp.jpy.util :as util]
+   [hkimjp.jpy.system :refer [start-system stop-system restart-system]]))
+
+;--------------------------
+(tel/set-min-level! :debug)
+
+(restart-system)
+
+; (start-system)
+; (stop-system)
+
+;---------------------------
+; clj-reload
+; hooks do not work?
+
+(defn before-ns-unload []
+  (println "called before-ns-unload"))
+
+(defn after-ns-reload []
+  (println "called after-ns-reload"))
 
 (reload/init
- {:dirs ["src" "dev" "test"]
-  :no-reload '#{user}
-  :unload-hook 'before-unload
-  :after-reload 'start-system})
-
-(defn before-unload []
-  (stop-system))
-
-(defn after-reload []
-  (start-system))
+ {:output :verbose
+  :dirs ["src" "test"]
+  :no-reload '#{user}})
 
 ; (reload/reload)
 
-(tel/set-min-level! :info)
+(defn reload []
+  (stop-system)
+  (reload/reload)
+  (start-system))
 
-(start-system)
+; (reload)
 
-; -------------------------
-(comment
-  ;; initialize
-  (ds/put! {:problem "dummy problem"
-            :stat "ok"
-            :datetime (jt/local-date-time)})
-  (ds/put! {:current 0})
-  ;;
-  (ds/qq '[:find ?e
-           :where
-           [?e _ _]])
-  (ds/pl 1)
-  (ds/pl 2)
-  (ds/pl 3)
-  (ds/pl 4)
-  (ds/pl 5)
-  :rcf)
-
-(comment
-  (ds/qq '[:find ?e ?num
-           :where
-           [?e :current ?num]])
-
-  (ds/pl 1)
-
-  (ds/qq '[:find ?problem
-           :where
-           [?e :num 19]
-           [?e :problem ?problem]])
-
-  (ds/qq '[:find ?e
-           :where
-           [?e _ _]])
-
-  (ds/pl 5)
-
-  :rcf)
-
+;-------------------