+++ src/hkimjp/jpy/help.clj
@@ -5,7 +5,8 @@
   (page
    [:div.m-4
     [:div.text-2xl.font-medium "授業中 Python"]
-    [:br]
+    [:p "workspace に表示する問題を自分のVScode/Python で解き、"
+     "回答を VScode からコピペでこちらに。"]
     [:div.font-bold "workspace"]
     [:div.m-4
      [:p "回答をアップロードする。過去回答のリストが見える。"]]
@@ -17,4 +18,4 @@
      [:p "新しくログインする。"]]
     [:div.font-bold "admin"]
     [:div.m-4
-     [:p "監視者専用ページ。問題を投入する。"]]]))
+     [:p "管理者専用ページ。問題を投入する。"]]]))