+++ src/hkimjp/jpy/login.clj
@@ -1,8 +1,9 @@
 (ns hkimjp.jpy.login
   (:require
    [buddy.hashers :as hashers]
+   [charred.api :as charred]
    [environ.core :refer [env]]
-   [hato.client :as hc]
+   [org.httpkit.client :as hk-client]
    [hiccup2.core :as h]
    [ring.util.anti-forgery :refer [anti-forgery-field]]
    [ring.util.response :as resp]
@@ -39,8 +40,12 @@
       (-> (resp/redirect "/workspace")
           (assoc-in [:session :identity] login)))
     (try
-      (let [resp (hc/get (str l22 "/api/user/" login) {:timeout 3000 :as :json})]
-        (if (and (some? resp) (hashers/check password (get-in resp [:body :password])))
+      (let [pw (-> (hk-client/get (str l22 "/api/user/" login))
+                   deref
+                   :body
+                   charred/read-json
+                   (get "password"))]
+        (if (hashers/check password pw)
           (do
             (t/log! :info (str "login success: " login))
             (-> (resp/redirect "/workspace")
@@ -49,7 +54,6 @@
             (t/log! :info (str "login failed: " login))
             (-> (resp/redirect "/")
                 (assoc :session {} :flash "login failed")))))
-      ;; maybe auth server error
       (catch Exception e
         (t/log! :warn (.getMessage e))
         (-> (resp/redirect "/")