+++ systemd/start.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+java --enable-native-access=ALL-UNNAMED -jar jpy.jar >> log/jpy.log 2>> log/error.log